Добавил возможность обновления чистой базы. Смотрите описания на вики для пользования классом. Теперь по-умолчанию, главная база убиваться не будет.

+++ trunk/src/java/ru/amse/agregator/quality/clusterization/ClusterizationProcess.java
@@ -16,45 +16,70 @@ import ru.amse.agregator.storage.UniqueId;
  */
 final public class ClusterizationProcess {
 
+    // indicates whether mainbase should be deleted
+    static private boolean shouldDeleteMainDB;
+    // dirty base name, should be 'dirty' by default
+    static private String baseToAddName;
+
     static public void perform() {
 
-        double threshold = 0.0;
-
-        System.out.println("Deleting main base");
+        // hard-coded thershold, we may choose to adjust this parameter later
+        final double threshold = 0.0;
 
         Database.connectToMainBase();
-        Database.removeAllCollections();
+        if (shouldDeleteMainDB) {
+            System.out.println("Deleting main base");
+            Database.removeAllCollections();
+        }
 
         ArrayList<String> types = DBWrapper.getTypeNames();
 
         for (String type : types) {
 
+            // instantiate all algorithms and structures used
             Clusterizer clusterizer = new PartitionClusterizer();
             Metric metric = new StandardMetric();
             Graph similatityGraph = new ArrayGraph(metric, threshold);
             ClusterStorage storage = new ArrayStorage();
 
             System.out.println("**************\n\nProcessing " + type + "s");
-            System.out.println("Retrieving data from dirty base");
+            System.out.println("Retrieving data from " + baseToAddName + " base");
 
-            Database.connectToDirtyBase();
+            // list of all objects to be processed
+            ArrayList<DBWrapper>  allObjects = new ArrayList<DBWrapper>();
+            
+            // prepare new objects
+            {
 
-            ArrayList<DBWrapper> allOfType = Database.getAllWithType(type);
-            setUniqueIdsForObjects(allOfType, Database.DIRTY_DB_NAME);
+                Database.connectToBase(baseToAddName);
+                ArrayList<DBWrapper> allOfTypeNew = Database.getAllWithType(type);
+                setUniqueIdsForObjects(allOfTypeNew, baseToAddName);
+                System.out.println("There are " + allOfTypeNew.size() + " " + type + "s in " + baseToAddName + " base");
+                allObjects.addAll(allOfTypeNew);
+            }
 
-            System.out.println("There are " + allOfType.size() + " " + type + "s in dirty base");
+            // prepare existing in main database objects objects
+            if (!shouldDeleteMainDB) {
+                //should merge with existing in main db objects
+                Database.connectToMainBase();
+                ArrayList<DBWrapper> allOfTypeOld = Database.getAllWithType(type);
+                setUniqueIdsForObjects(allOfTypeOld, Database.MAIN_DB_NAME);
+                System.out.println("There are " + allOfTypeOld.size() + " " + type + "'s in main base");
+                allObjects.addAll(allOfTypeOld);
+            }
 
+        
             System.out.println("Building similarity graph\nObjects processed:");
-            similatityGraph.build(allOfType);
+            similatityGraph.build(allObjects);
 
-            System.out.println(String.valueOf(allOfType.size()));
+            System.out.println(String.valueOf(allObjects.size()));
 
             System.out.println("Resulting graph has "
                     + String.valueOf(similatityGraph.getEdgeCount()) + " edges");
 
             System.out.println("Performing clusterization process");
 
-            clusterizer.clusterize(allOfType, similatityGraph, storage);
+            clusterizer.clusterize(allObjects, similatityGraph, storage);
             
             //print non single object clusters
 //
@@ -73,11 +98,11 @@ final public class ClusterizationProcess {
 //            }
 //            storage.finishIterating();
 
-            System.out.println(String.valueOf(allOfType.size()));
+            System.out.println(String.valueOf(allObjects.size()));
             
             System.out.println("Clusterisation process created "
                     + String.valueOf(storage.getClusterCount()) + " clusters out of "
-                    + String.valueOf(allOfType.size()) + " objects");
+                    + String.valueOf(allObjects.size()) + " objects");
 
             System.out.println("Merging and adding to main base");
 
@@ -96,6 +121,36 @@ final public class ClusterizationProcess {
     }
 
     public static void main(String[] args) {
+        //set default parameter values
+        baseToAddName = Database.DIRTY_DB_NAME;
+        shouldDeleteMainDB = false;
+
+        if (args.length > 2) {
+            System.out.println("Error: Unexpected number of arguments");
+            return;
+        }
+        if (args.length >= 1) {
+            baseToAddName = args[0];           
+        }
+
+        if (args.length == 2) {
+            if (args[1].equals("-r")) {
+                shouldDeleteMainDB = true;
+            } else if (args[1].equals("-u")) {
+                shouldDeleteMainDB = false;
+            } else {
+                System.out.println("Error: Unexpected second argument");
+                return;
+            }
+        }
+
+        System.out.println("Adding " + baseToAddName + " database");
+        if (shouldDeleteMainDB) {
+            System.out.println("Renewing database. Main database will be deleted");
+        } else {
+            System.out.println("Updating database. New objects will be added");
+        }
+
         perform();
     }
 }
